refactor: Consolidate Drawer, Header, & Nav into a single component
- Abstract logic from `App.tsx`
